cleanup
git-svn-id: https://only.mawhrin.net/repos/stuff/trunk@288 62a0de91-87f2-0310-aeaa-8eaa5733e38e

+++ fbreader/common/textview/ViewStyle.cpp
@@ -90,7 +90,7 @@ int TextView::ViewStyle::elementWidth(const WordCursor &cursor) const {
 		case TextElement::EMPTY_LINE_ELEMENT:
 			return context().width() + abs(style().leftIndent()) + abs(style().rightIndent()) + abs(style().firstLineIndentDelta() + 1);
 		case TextElement::TREE_ELEMENT:
-			return (int)(context().stringHeight() * 4 / 3);
+			return context().stringHeight() * 4 / 3;
 		case TextElement::CONTROL_ELEMENT:
 			return 0;
 	}
